+++ nope/lib/nope/Nope/DateTime.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Nope;
+
+class DateTime extends \Carbon\Carbon implements \JsonSerializable {
+
+  function jsonSerialize() {
+    return $this->__toString();
+  }
+
+}