+++ nope/lib/queries/Page.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Nope\Query;
+
+class Page extends \Nope\Query {
+
+}