+++ nope/lib/register/models.php
@@ -14,14 +14,20 @@
   ]
 ]);
 \Nope::registerModel('setting', [
-  'model' => NOPE_LIB_DIR . 'models/Setting.php',
+  'model' => [
+    NOPE_LIB_DIR . 'models/Setting.php',
+    NOPE_LIB_DIR . 'queries/Setting.php'
+  ],
   'route' => NOPE_LIB_DIR . 'routes/setting.php',
   'js' => [
     'lib/assets/js/setting.js'
   ]
 ]);
 \Nope::registerModel('page', [
-  'model' => NOPE_LIB_DIR . 'models/Page.php',
+  'model' => [
+    NOPE_LIB_DIR . 'models/Page.php',
+    NOPE_LIB_DIR . 'queries/Page.php'
+  ],
   'route' => NOPE_LIB_DIR . 'routes/page.php'
 ]);
 \Nope::registerModel('media', [