+++ nope/theme/default/_common/header.php
@@ -1,0 +1,11 @@
+<!doctype html>
+<html lang="en">
+<head>
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <link rel="stylesheet" href="http://fonts.googleapis.com/css?family=Roboto:300,300italic,700,700italic">
+  <link rel="stylesheet" href="<?php echo asset('assets/css/normalize.css'); ?>">
+  <link rel="stylesheet" href="<?php echo asset('assets/css/milligram.min.css'); ?>">
+  <link rel="stylesheet" href="<?php echo asset('assets/css/style.css'); ?>">
+</head>
+<body>