+++ nope/theme/default/page.php
@@ -1,5 +1,11 @@
-<h2><?php echo $content->title; ?></h2>
-
-<div>
-  <?php echo $content->body; ?>
-</div>
+<?php include NOPE_THEME_DIR . '_common/header.php'; ?>
+<main class="wrapper">
+  <section class="container">
+    <article>
+      <h5><?php echo $content->title; ?></h5>
+      <h6>Published <?php echo $content->startPublishingDate->diffForHumans(); ?></h6>
+      <?php echo $content->body; ?>
+    </article>
+  </section>
+</main>
+<?php include NOPE_THEME_DIR . '_common/footer.php'; ?>