+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ### Libraries
 - Updated **angular** to version 1.4.9.
 - Updated `composer` libraries.
+- Added **Carbon** version 1.21.
+
+### Features
+- Added `Nope\String`, `Nope\Text`, `Nope\DateTime` classes.
+- Added `Nope\Query` class to obtain items different from JSON-encoded data.
+- Default theme based on [milligram](https://github.com/milligram/milligram).
 
 ## v0.8.0 (2016-01-27)